+++ Code/Commando/dlgmplanhostoptions.cpp
@@ -1225,7 +1225,7 @@ MPLanHostMapCycleOptionsTabClass::Build_Mod_Package_List (void)
 
 			int item_index = combobx_ctrl->Add_String (curr_name);
 			if (item_index != -1) {
-				combobx_ctrl->Set_Item_Data (item_index, (uint32)package);
+				combobx_ctrl->Set_Item_Data (item_index, (uintptr_t)package);
 
 				//
 				//	Is this the default entry?  If so select it...
